+++ lib/less/rails/bootstrap/engine.rb
@@ -4,7 +4,7 @@ module Less
       class Engine < ::Rails::Engine
         
         initializer 'less-rails-bootstrap.setup', :after => 'less-rails.after.load_config_initializers', :group => :all do |app|
-          app.config.less.paths << File.join(config.root, 'vendor', 'assets', 'stylesheets')
+          app.config.less.paths << File.join(config.root, 'vendor', 'frameworks')
         end
         
       end